test-task-224/pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["pavivin <pavivin@yandex.ru>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = {extras = ["asyncpg"], version = "^2.0.9"}
httpx = "^0.24.0"
orjson = "^3.8.10"
pydantic = {extras = ["email"], version = "^1.10.7"}
alembic = "^1.10.3"
pytest-asyncio = "^0.21.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
factory-boy = "^3.2.1"
async-factory-boy = "^1.0.0"
psycopg2-binary = "^2.9.6"
aiohttp = "^3.8.4"
fastapi = "0.99.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"
pytest-factoryboy = "^2.5.1"
punq = "^0.6.2"
shapely = "^2.0.1"
nltk = "^3.8.1"
aiofiles = "^23.1.0"
pillow = "^10.0.0"
firebase-admin = "^6.2.0"
uuid7 = "^0.1.0"
async-lru = "^2.0.4"
python-multipart = "^0.0.9"
sentry-sdk = {extras = ["celery"], version = "^1.38.0"}
phonenumbers = "^8.13.22"
aiobotocore = "^2.7.0"
webp = "^0.3.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
redis = "^5.0.4"
taskiq-aio-pika = "^0.4.0"
taskiq-redis = "^0.5.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "venv",
    "migrations"
]

line-length = 120

target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]

quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"
